requirements.txt
apache-airflow-providers-oracle[common.sql]